Update README.md
The avatar movie is the one that I want is not totally my favorite but I like that movie like Cameron said, 'Avatar asks us to see that everything is connected, all human beings to each other, and us to the Earth. ' This is a clear statement of religious belief. This is pantheism.